# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/message/file": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "发布文件类型的消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "关系id",
                        "name": "relation_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "回复消息id",
                        "name": "rly_msg_id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamCreateFileMsg"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/message/info/top": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "获取指定关系中的置顶消息，如果不存在则为 null",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "relation_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "完整的消息详情,但不存在则为null",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamGetTopMsgByRelationID"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/message/list/content": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "通过内容模糊查找指定或所有关系中的消息，按照时间先后顺序倒序排序，不会查询撤回的消息（指定关系 ID \u003c 0 则查询所有关系中的消息）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "content",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "第几页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页的大小",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "relation_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "完整的消息详情，包含回复消息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamGetMsgsByContent"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/message/list/offer": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "获取所有关系指定时间戳之后的信息，获取的消息按照发布时间先后排序，同时包含是否已读的标识",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "拉取消息的最晚时间戳（精确到秒）",
                        "name": "last_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "第几页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页的大小",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "完整的消息详情，包含回复消息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamOfferMsgsByAccountIDAndTime"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/message/list/pin": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "获取指定关系的 pin 消息，按照pin时间倒序排序",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "第几页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页的大小",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "relation_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "完整的消息详情，包含回复消息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamGetPinMsgsByRelationID"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/message/list/reply": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "获取指定消息的所有回复消息，按照回复时间先后排序",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "msg_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "第几页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页的大小",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "relation_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "完整的消息详情，包含回复消息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamGetRlyMsgsInfoByMsgID"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/message/list/time": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "获取指定关系指定时间戳之前的信息，获取的消息按照发布时间先后排序",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "拉取消息最晚的时间戳（精确到秒）",
                        "name": "last_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "第几页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页的大小",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "关系 ID",
                        "name": "relation_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "完整的消息详情，包含回复消息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamGetMsgsRelationIDAndTime"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/message/update/pin": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "更改消息的 pin 状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "消息 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否 pin",
                        "name": "is_pin",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "关系 ID",
                        "name": "relation_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 2010:账号不存在 5001:消息不存在",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/message/update/revoke": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "撤回消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "消息 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 2010:账号不存在 5001:消息不存在",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/message/update/top": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "更改消息的置顶状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "消息 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否置顶",
                        "name": "is_top",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "关系 ID",
                        "name": "relation_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 2010:账号不存在 5001:消息不存在",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.State": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码，0：成功；否则失败",
                    "type": "integer"
                },
                "data": {
                    "description": "数据，失败时返回空"
                },
                "msg": {
                    "description": "状态的具体描述",
                    "type": "string"
                }
            }
        },
...
        "reply.ParamCreateFileMsg": {
            "type": "object",
            "properties": {
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "file_id": {
                    "description": "文件 ID",
                    "type": "integer"
                },
                "id": {
                    "description": "消息 ID",
                    "type": "integer"
                },
                "msg_content": {
                    "description": "消息内容，文件则为 url",
                    "type": "string"
                }
            }
        },
...
        "reply.ParamGetMsgsByContent": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reply.ParamBriefMsgInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "reply.ParamGetTopMsgByRelationID": {
            "type": "object",
            "properties": {
                "msg_info": {
                    "description": "置顶消息详情",
                    "allOf": [
                        {
                            "$ref": "#/definitions/reply.ParamMsgInfo"
                        }
                    ]
                }
            }
        },
...
        "reply.ParamOfferMsgsByAccountIDAndTime": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reply.ParamMsgInfoWithRlyAndHasRead"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}